Rename resources module to cwxml
