+++ cwxml/clipsdictionary.py
