+++ cwxml/drawable.py
