+++ ybn/ybnexport.py
@@ -1,5 +1,5 @@
 from .properties import CollisionMatFlags
-from ..resources.bound import *
+from ..cwxml.bound import *
 from ..sollumz_properties import BOUND_SHAPE_TYPES, MaterialType, SollumType
 from ..tools.meshhelper import *
 from ..tools.utils import *
