+++ ydr/shader_materials.py
@@ -1,7 +1,7 @@
 import bpy
 
 from ..tools.version import USE_LEGACY
-from ..resources.shader import ShaderManager
+from ..cwxml.shader import ShaderManager
 from ..sollumz_properties import MaterialType
 from collections import namedtuple
