+++ ynv/ynvimport.py
@@ -1,5 +1,5 @@
 from ..tools.meshhelper import create_box
-from ..resources.navmesh import YNV
+from ..cwxml.navmesh import YNV
 from ..sollumz_properties import SOLLUMZ_UI_NAMES, SollumType
 import os
 import bpy
